// components-vue/src/Essentials/Input/Calendar/Calendar.types.ts
export type CalendarDayRenderer = (props: {
    day: DateValue;
    flags: InteractionFlags<CalendarRenderProps>;
}) => VNodeChild;

export type CalendarWeekdayRenderer = (props: { name: string; index: number }) => VNodeChild;
// ...
export type CalendarSlots = {
    /** Draws one day cell. */
    renderDay: CalendarDayRenderer;
    /** Draws one weekday heading. */
    renderWeekday?: CalendarWeekdayRenderer;
};